Refactor: App 클래스와 BaseballGame 클래스 분리
- App: 게임 시작 부분
- BaseballGame: 게임 동작 부분
