+++ docs/README.md
@@ -5,7 +5,7 @@
   - 서로 다른 수
   - mission-utils 라이브러리의 random 사용
 
-- [x] `숫자 야구 게임을 시작합니다.` 출력하기
+- [x] `숫자 야구 게임을 시작합니다.` 출력하기 (가장 첫 게임에서만 출력)
 
 - [x] 사용자에게 서로다른 3자리 수 입력받기
 ```
